refactor:优化DDD架构

+++ AutoSimpleAgent/src/main/java/org/feynix/application/agent/service/AgentWorkspaceAppService.java
@@ -1,15 +1,22 @@
 package org.feynix.application.agent.service;
 
 
+import org.feynix.application.agent.assembler.AgentAssembler;
 import org.feynix.application.agent.dto.AgentDTO;
+import org.feynix.domain.agent.model.AgentEntity;
+import org.feynix.domain.agent.model.AgentWorkspaceEntity;
 import org.feynix.domain.agent.service.AgentDomainService;
 import org.feynix.domain.agent.service.AgentWorkspaceDomainService;
+import org.feynix.domain.conversation.model.SessionEntity;
 import org.feynix.domain.conversation.service.ConversationDomainService;
 import org.feynix.domain.conversation.service.SessionDomainService;
 import org.feynix.application.conversation.dto.SessionDTO;
+import org.feynix.domain.llm.model.ModelEntity;
+import org.feynix.domain.llm.service.LLMDomainService;
 import org.feynix.infrastructure.exception.BusinessException;
 import org.feynix.interfaces.dto.agent.SearchAgentsRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -31,13 +38,15 @@ public class AgentWorkspaceAppService {
     private final SessionDomainService sessionDomainService;
 
     private final ConversationDomainService conversationDomainService;
+    private final LLMDomainService llmDomainService;
 
     public AgentWorkspaceAppService(AgentWorkspaceDomainService agentWorkspaceDomainService,
-                                    AgentDomainService agentServiceDomainService, SessionDomainService sessionDomainService, ConversationDomainService conversationDomainService) {
+                                    AgentDomainService agentServiceDomainService, SessionDomainService sessionDomainService, ConversationDomainService conversationDomainService,LLMDomainService llmDomainService) {
         this.agentWorkspaceDomainService = agentWorkspaceDomainService;
         this.agentServiceDomainService = agentServiceDomainService;
         this.sessionDomainService = sessionDomainService;
         this.conversationDomainService = conversationDomainService;
+        this.llmDomainService = llmDomainService;
     }
 
 
@@ -48,15 +57,8 @@ public class AgentWorkspaceAppService {
      * @return
      */
     public List<AgentDTO> getAgents(String userId) {
-        // 1.获取当前用户的所有助理
-        List<AgentDTO> userAents = agentServiceDomainService.getUserAgents(userId, new SearchAgentsRequest());
-
-        // 2.获取已添加到工作区的助理
-        List<AgentDTO> workspaceAgents = agentWorkspaceDomainService.getWorkspaceAgents(userId);
-
-        // 合并两个列表
-        userAents.addAll(workspaceAgents);
-        return userAents;
+        List<AgentEntity> workspaceAgents = agentWorkspaceDomainService.getWorkspaceAgents(userId);
+        return AgentAssembler.toDTOs(workspaceAgents);
     }
 
     /**
@@ -64,20 +66,52 @@ public class AgentWorkspaceAppService {
      * @param agentId 助理id
      * @param userId 用户id
      */
+    @Transactional
     public void deleteAgent(String agentId, String userId) {
+
+        // agent如果是自己的则不允许删除
+        AgentEntity agent = agentServiceDomainService.getAgentById(agentId);
+        if (agent.getUserId().equals(userId)){
+            throw new BusinessException("该助理属于自己，不允许删除");
+        }
+
         boolean deleteAgent = agentWorkspaceDomainService.deleteAgent(agentId, userId);
         if (!deleteAgent){
             throw new BusinessException("删除助理失败");
         }
-        // 查出会话列表,收集 sessionIds
-
-        List<String> sessionIds = sessionDomainService.getSessionsByAgentId(agentId).stream().map(SessionDTO::getId).collect(Collectors.toList());
+        List<String> sessionIds = sessionDomainService.getSessionsByAgentId(agentId).stream().map(SessionEntity::getId).collect(Collectors.toList());
         if (sessionIds.isEmpty()){
             return;
         }
         sessionDomainService.deleteSessions(sessionIds);
-        // 删除agent下的会话
-        // 删除会话下的所有消息
         conversationDomainService.deleteConversationMessages(sessionIds);
+    }
+
+    /**
+     * 保存模型
+     * @param agentId agentId
+     * @param userId 用户id
+     * @param modelId 模型id
+     */
+    public void saveModel(String agentId, String userId, String modelId) {
+
+        // 模型是否是自己的 or 官方的
+        ModelEntity model = llmDomainService.getModelById(modelId);
+        if (!model.getOfficial() && !model.getUserId().equals(userId)) {
+            throw new BusinessException("模型不存在");
+        }
+
+        AgentWorkspaceEntity workspace = agentWorkspaceDomainService.findWorkspace(agentId, userId);
+        if (workspace == null){
+            workspace = new AgentWorkspaceEntity();
+            workspace.setAgentId(agentId);
+            workspace.setUserId(userId);
+        }
+        workspace.setModelId(modelId);
+        agentWorkspaceDomainService.save(workspace);
+    }
+
+    public String getConfiguredModelId(String agentId, String userId) {
+        return agentWorkspaceDomainService.getWorkspace(agentId,userId).getModelId();
     }
 }
